requirements.txt
#Third party package requirements for OpenEats 2
akismet==0.2.0
beautifulsoup4==4.5.1
coreapi-cli
Django==1.10
django-appconf==1.0.2
django-extensions==1.7.2
django-grappelli==2.8.1
django-haystack==2.5.0
django-imagekit==3.3
django-pagination==1.0.7
django-registration==2.1.2
django-reversion==1.5.1
django-taggit==0.21.1
django-tastypie==0.13.3
django-templatetag-sugar==1.0
django-webtest==1.7.5
djangorestframework==3.4.6
feedparser==5.1.3
gunicorn==19.6.0
html5lib==0.95
microsofttranslator==0.8
pilkit==1.1.13
Pillow==2.7.0
polib==1.0.7
Pygments==2.1.3
python-dateutil==2.5.3
python-mimeparse==1.5.2
reportlab==2.5
requests==2.11.1
six==1.10.0
sqlparse==0.2.1
waitress==0.9.0
WebOb==1.6.1
WebTest==2.0.7
Whoosh==2.5.2